requirements.txt
transformers>=4.36.0
diffusers>=0.26.0
accelerate>=0.21.0
peft>=0.6.0
Pillow>=9.5.0
numpy>=1.23.0
scipy>=1.10.0